Modify migrations

+++ db/migrate/20170817181852_remove_mealkind_from_meals.rb
@@ -1,5 +1,7 @@
 class RemoveMealkindFromMeals < ActiveRecord::Migration[5.1]
   def change
+  	if column_exists? :meals, :mealkind
 	    remove_column :meals, :mealkind, :string
+	  end
   end
 end
